'photo'

+++ review.php
@@ -32,6 +32,7 @@ $arr=mysqli_fetch_array($res);
 $roomid=$arr['roomid'];
 
 $d=$_POST["rev"] ;
+echo $d ;
 $roomno="";
     $hostelno="";
     
@@ -49,6 +50,7 @@ $roomno="";
     $ar=mysqli_fetch_array($add);
     $ri=$ar['max(rid)'] ;
     $rid=$ri+1 ;
+
     $a="insert into reviews(roomid,description,rid,photo) values('$roomid','$d',$rid,'$folder')" ;
 mysqli_query($con,$a) ;
     $b="insert into writes_rev(rid,sid) values($rid,$m)" ;
@@ -167,11 +169,19 @@ $q="select * from student ,reviews,writes_rev where student.sid='$m' and writes_
 $res=mysqli_query($con,$q);
 $arr=mysqli_fetch_array($res);
 
-  echo $arr["roomid"] ;?></h2> </div>
-  <div class="pic"><?php echo "<p><a href='".$arr['photo']."'><img src='".$arr['photo']."' width='200'  height='200' ></a></p> ";?></div>
-    <em>
+  echo $arr["roomid"] ;?></h2> </div><em>
+ <?php 
+ if(!$arr['photo']||$arr['description']){
+ if(!$arr['photo']){
+   echo "<p align=center><a href='".$arr['photo']."'><img src='".$arr['photo']."' width='200'  height='200' ></a></p> ";}
+    
    
-  <br><br><br><?php echo $arr["description"] ; ?><br>By<br><?php  echo $_SESSION['sid']  ; ?><br></em></div>
+  echo "<br><br><br>".$arr["description"]."<br>By<br>".$_SESSION['sid']  ;}
+
+  else
+     echo "No reviews yet" ;
+   ?><br></em></div>
+
   <a href="delete.php" class="btn btn-a">Delete</a>
   </div>
